offload_percentage should be float

+++ vllm/entrypoints/openai/chat_completion/protocol.py
@@ -203,7 +203,7 @@ class ChatCompletionRequest(OpenAIBaseModel):
     prompt_logprobs: int | None = None
     allowed_token_ids: list[int] | None = None
     bad_words: list[str] = Field(default_factory=list)
-    offload_prompt_percentage: int | None = 100
+    offload_prompt_percentage: float | None = 100
     # --8<-- [end:chat-completion-sampling-params]
 
     # --8<-- [start:chat-completion-extra-params]
